requirements.txt
django==6.0.5
djangorestframework==3.17.1
Pillow==12.2.0
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
